componentize list of cards
